@@ -1,16 +1,17 @@
 {
 	"name": "camelcase",
-	"version": "5.3.1",
+	"version": "6.3.0",
 	"description": "Convert a dash/dot/underscore/space separated string to camelCase or PascalCase: `foo-bar` → `fooBar`",
 	"license": "MIT",
 	"repository": "sindresorhus/camelcase",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
 	"engines": {
-		"node": ">=6"
+		"node": ">=10"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -37,7 +38,7 @@
 	],
 	"devDependencies": {
 		"ava": "^1.4.1",
-		"tsd": "^0.7.1",
-		"xo": "^0.24.0"
+		"tsd": "^0.11.0",
+		"xo": "^0.28.3"
 	}
 }
