@@ -1,20 +1,17 @@
 {
 	"name": "escape-string-regexp",
-	"version": "2.0.0",
+	"version": "4.0.0",
 	"description": "Escape RegExp special characters",
 	"license": "MIT",
 	"repository": "sindresorhus/escape-string-regexp",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
-	"maintainers": [
-		"Sindre Sorhus <sindresorhus@gmail.com> (sindresorhus.com)",
-		"Joshua Boy Nicolai Appelman <joshua@jbna.nl> (jbna.nl)"
-	],
 	"engines": {
-		"node": ">=8"
+		"node": ">=10"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -27,17 +24,15 @@
 		"escape",
 		"regex",
 		"regexp",
-		"re",
 		"regular",
 		"expression",
 		"string",
-		"str",
 		"special",
 		"characters"
 	],
 	"devDependencies": {
 		"ava": "^1.4.1",
-		"tsd": "^0.7.2",
-		"xo": "^0.24.0"
+		"tsd": "^0.11.0",
+		"xo": "^0.28.3"
 	}
 }
