@@ -1,16 +1,17 @@
 {
 	"name": "find-up",
-	"version": "4.1.0",
+	"version": "5.0.0",
 	"description": "Find a file or directory by walking up parent directories",
 	"license": "MIT",
 	"repository": "sindresorhus/find-up",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
 	"engines": {
-		"node": ">=8"
+		"node": ">=10"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -40,14 +41,14 @@
 		"path"
 	],
 	"dependencies": {
-		"locate-path": "^5.0.0",
+		"locate-path": "^6.0.0",
 		"path-exists": "^4.0.0"
 	},
 	"devDependencies": {
 		"ava": "^2.1.0",
 		"is-path-inside": "^2.1.0",
-		"tempy": "^0.3.0",
-		"tsd": "^0.7.3",
-		"xo": "^0.24.0"
+		"tempy": "^0.6.0",
+		"tsd": "^0.13.1",
+		"xo": "^0.33.0"
 	}
 }
