@@ -1,16 +1,17 @@
 {
 	"name": "p-locate",
-	"version": "4.1.0",
+	"version": "5.0.0",
 	"description": "Get the first fulfilled promise that satisfies the provided testing function",
 	"license": "MIT",
 	"repository": "sindresorhus/p-locate",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
 	"engines": {
-		"node": ">=8"
+		"node": ">=10"
 	},
 	"scripts": {
 		"test": "xo && ava && tsd"
@@ -40,14 +41,14 @@
 		"bluebird"
 	],
 	"dependencies": {
-		"p-limit": "^2.2.0"
+		"p-limit": "^3.0.2"
 	},
 	"devDependencies": {
-		"ava": "^1.4.1",
+		"ava": "^2.4.0",
 		"delay": "^4.1.0",
-		"in-range": "^1.0.0",
-		"time-span": "^3.0.0",
-		"tsd": "^0.7.2",
-		"xo": "^0.24.0"
+		"in-range": "^2.0.0",
+		"time-span": "^4.0.0",
+		"tsd": "^0.13.1",
+		"xo": "^0.32.1"
 	}
 }
