@@ -1,8 +1,10 @@
-import mongoose from 'mongoose';
+
+
+import mongoose from "mongoose";
 
 const reviewSchema = new mongoose.Schema({
   review_id: { type: String, required: true },
-  reviewed_van_id: { type: String, required: true, ref: 'Van' },
+  reviewed_van_id: { type: String, required: true, ref: "Van" },
   review_date: { type: Date, default: Date.now },
   review_title: { type: String, required: true },
   review_stars: { type: Number, required: true, min: 1, max: 5 },
@@ -13,6 +15,7 @@ const reviewSchema = new mongoose.Schema({
     recommendation: { type: Boolean, required: true },
   },
   review_source: String,
+  
 });
 
-export default mongoose.model('Review', reviewSchema);
+export default mongoose.model("Review", reviewSchema);
